backend/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
axios==0.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
filelock==3.20.1
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.9.23
fsspec==2025.12.0
ftfy==6.3.1
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.13.0
libclang==18.1.1
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numpy==2.3.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==12.0.0
protobuf==6.33.2
Pygments==2.19.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.8.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.2
wcwidth==0.2.14
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
